fix: edit code RegistrationForm, EntryDataForm, registratForm, requestAPI
